Refactor project card
