+++ web/src/components/molecules/Dashboard/ProjectList.tsx
@@ -1,14 +1,9 @@
-import {
-  EditOutlined,
-  EllipsisOutlined,
-  PlusOutlined,
-  SettingOutlined,
-} from "@ant-design/icons";
+import { PlusOutlined } from "@ant-design/icons";
 import styled from "@emotion/styled";
+import Button from "@reearth-cms/components/atoms/Button";
+import ProjectCard from "@reearth-cms/components/molecules/Dashboard/ProjectCard";
 import { Project as ProjectType } from "@reearth-cms/components/molecules/Dashboard/types";
-import { Button, Card } from "antd";
 import React from "react";
-import { useNavigate } from "react-router-dom";
 
 export interface Props {
   className?: string;
@@ -23,30 +18,16 @@ const ProjectList: React.FC<Props> = ({
   workspaceId,
   handleProjectModalOpen,
 }) => {
-  const navigate = useNavigate();
-  const { Meta } = Card;
   return (
     <StyledDashboardBlock className={className}>
       {projects?.length ? (
         <Content>
-          {projects.map((project) => (
-            <CardWrapper key={project.id}>
-              <Card
-                cover={<Cover>{project.name.charAt(0)}</Cover>}
-                actions={[
-                  <SettingOutlined
-                    onClick={() =>
-                      navigate("/workspaces/" + workspaceId + "/" + project.id)
-                    }
-                    key="setting"
-                  />,
-                  <EditOutlined key="edit" />,
-                  <EllipsisOutlined key="ellipsis" />,
-                ]}
-              >
-                <Meta title={project.name} description={project.description} />
-              </Card>
-            </CardWrapper>
+          {projects.map((project: ProjectType) => (
+            <ProjectCard
+              key={project.id}
+              project={project}
+              workspaceId={workspaceId}
+            ></ProjectCard>
           ))}
         </Content>
       ) : (
@@ -103,11 +84,6 @@ const StyledDashboardBlock = styled.div`
   }
 `;
 
-const CardWrapper = styled.div`
-  width: 300px;
-  padding: 12px;
-`;
-
 const Content = styled.div`
   height: 100%;
   margin: 0 -12px;
@@ -115,18 +91,6 @@ const Content = styled.div`
   flex-wrap: wrap;
   justify-content: flex-start;
   align-content: flex-start;
-`;
-
-const Cover = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-weight: 500;
-  font-size: 38px;
-  line-height: 46px;
-  height: 150px;
-  background-color: #eeeeee;
-  color: #fff;
 `;
 
 export default ProjectList;
